+++ samples/CG.Business.QuickStart/Startup.cs
@@ -2,9 +2,15 @@
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using System;
 
 namespace CG.Business.QuickStart
 {
+    // This code demonstrates how, through the AddRepositories/UseRepositories, and
+    //   AddStrategies/UseStrategies methods, along with corresponding JSON in the
+    //   appSettings.json, we can dynamically load & configure repository and strategy
+    //   types.
+
     public class Startup
     {
         public Startup(IConfiguration configuration)
@@ -17,22 +23,43 @@ namespace CG.Business.QuickStart
         public void ConfigureServices(IServiceCollection services)
         {
             // This is how to register repositories, using the configuration
-            //   to control what repository types get registered, and what
+            //   to control what repository type(s) get registered, and what
             //   options are passed to those repositories, at runtime.
             services.AddRepositories(
-                Configuration.GetSection("Foo")
+                Configuration.GetSection("FooRep")
                 );
+
+            Console.WriteLine();
+
+            // This is how to register strategies, using the configuration
+            //   to control what repository type(s) get registered, and what
+            //   options are passed to those strategies, at runtime.
+            services.AddStrategies(
+                Configuration.GetSection("FooStrat")
+                );
+
+            Console.WriteLine();
         }
 
         public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             // This is how to call any startup logic required for the configured
-            //   repository types - for instance, EFCore types might need to apply
-            //   migrations.
+            //   repository types.
             app.UseRepositories(
                 env,
-                Configuration.GetSection("Foo")
+                Configuration.GetSection("FooRep")
                 );
+
+            Console.WriteLine();
+
+            // This is how to call any startup logic required for the configured
+            //   strategy types.
+            app.UseStrategies(
+                env,
+                Configuration.GetSection("FooStrat")
+                );
+
+            Console.WriteLine();
         }
     }
 }