+++ src/CG.Business/Properties/Resources.Designer.cs
@@ -61,6 +61,15 @@ namespace CG.Business.Properties {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to {0} error! The method failed to process one or more selections! section: &apos;{1}&apos;..
+        /// </summary>
+        internal static string AggregatedFail {
+            get {
+                return ResourceManager.GetString("AggregatedFail", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to {0} error! The method encountered an empty or missing &apos;Name&apos; property!.
         /// </summary>
         internal static string EmptyStrategyName {
@@ -88,6 +97,15 @@ namespace CG.Business.Properties {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to {0} error! The method encountered a missing or malformed &apos;Selection&apos; property! section: &apos;{1}&apos;..
+        /// </summary>
+        internal static string MalformedSelectionArray {
+            get {
+                return ResourceManager.GetString("MalformedSelectionArray", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to {0} error! Failed to locate an extension method named: &apos;{1}&apos;, parameters: &apos;{2}&apos;!.
         /// </summary>
         internal static string MethodNotFound {
@@ -97,11 +115,29 @@ namespace CG.Business.Properties {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to {0} error! The method was pointed to: &apos;{1}&apos; but no &apos;Selected&apos; field was found there!.
+        /// </summary>
+        internal static string MissingSelectedField {
+            get {
+                return ResourceManager.GetString("MissingSelectedField", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to Unable to load the specified assembly: &apos;{0}&apos;! See inner exception(s) for more detail. .
         /// </summary>
         internal static string NoLoadAssembly {
             get {
                 return ResourceManager.GetString("NoLoadAssembly", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to {0} error! The method failed to process selection: &apos;{1}&apos;! section: &apos;{2}&apos;..
+        /// </summary>
+        internal static string SingleFail {
+            get {
+                return ResourceManager.GetString("SingleFail", resourceCulture);
             }
         }
     }